- Totally rework static plugin - Remove old one, now it is part of the HTTP plugin
