+++ tests/plugins/http/configs/.rr-http-static-disabled.yaml
@@ -8,19 +8,15 @@ server:
   relay_timeout: "20s"
 
 http:
-  address: 127.0.0.1:21603
+  address: 127.0.0.1:21234
   max_request_size: 1024
-  middleware: [ "gzip", "static" ]
+  middleware: [ "gzip" ]
   trusted_subnets: [ "10.0.0.0/8", "127.0.0.0/8", "172.16.0.0/12", "192.168.0.0/16", "::1/128", "fc00::/7", "fe80::/10" ]
   uploads:
     forbid: [ ".php", ".exe", ".bat" ]
   static:
-    dir: "../../../tests"
-    forbid: [ "" ]
-    request:
-      "input": "custom-header"
-    response:
-      "output": "output-header"
+    dir: "abc" #not exists
+    forbid: [ ".php", ".htaccess" ]
   pool:
     num_workers: 2
     max_jobs: 0
