+++ tests/plugins/http/configs/.rr-http-static-files.yaml
@@ -10,14 +10,15 @@ server:
 http:
   address: 127.0.0.1:34653
   max_request_size: 1024
-  middleware: [ "gzip", "static" ]
+  middleware: [ "gzip" ]
   trusted_subnets: [ "10.0.0.0/8", "127.0.0.0/8", "172.16.0.0/12", "192.168.0.0/16", "::1/128", "fc00::/7", "fe80::/10" ]
   uploads:
     forbid: [ ".php", ".exe", ".bat" ]
   static:
-    dir: "../../../tests"
+    dir: "../../../"
+    pattern: "/tests/"
+    allow: [ ".ico" ]
     forbid: [ ".php", ".htaccess" ]
-    always: [ ".ico" ]
 
   pool:
     num_workers: 2
